General cleanup
Move to VS2017 RC2 project files
Get tests running again
Fix the failing test (change the behavior of None operations in the C
style evaluator to not perform a type conversion on the scope - in
particular, not converting it to a boolean)
Move conditional test regions in to separate files
Make versions consistent across projects (version.props)
Delete unnecessary AssemblyInfo files
