+++ src/dotnet-new3/dotnet-new3.csproj
@@ -1,21 +1,10 @@
-﻿<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" />
-  
+﻿<Project ToolsVersion="15.0" Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\Version.props" />
   <PropertyGroup>
     <Description>Template Instantiation Commands for .NET Core CLI.</Description>
-    <TargetFramework>netcoreapp1.0</TargetFramework>
-    <TargetFrameworkVersion Condition=" '$(TargetFramework)' != 'netcoreapp1.0' ">v4.5.1</TargetFrameworkVersion>
-    <WarningsAsErrors>true</WarningsAsErrors>
-    <AssemblyName>dotnet-new3</AssemblyName>
-    <RuntimeIdentifiers>win7-x64;win7-x86;osx.10.10-x64;osx.10.11-x64;ubuntu.14.04-x64;ubuntu.16.04-x64;centos.7-x64;rhel.7.2-x64;debian.8-x64;fedora.23-x64;opensuse.13.2-x64</RuntimeIdentifiers>
-    <PackageVersion>1.0.0-beta1</PackageVersion>
+    <TargetFrameworks>netcoreapp1.0</TargetFrameworks>
     <OutputType>Exe</OutputType>
-    <PackageTargetFallback>$(PackageTargetFallback);portable-dnxcore50+net45+win8+wp8+wpa81</PackageTargetFallback>
-    <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
-    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
-    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
-    <OutputPath Condition="'$(OutputPath)'=='' ">bin\$(Configuration)</OutputPath>
-	<TargetFrameworkFallback>$(TargetFrameworkFallback);netstandard1.1;netstandard1.2;netstandard1.3;netstandard1.4;netstandard1.5;netstandard1.6</TargetFrameworkFallback>
+    <RuntimeIdentifiers>win7-x64;win7-x86;osx.10.10-x64;osx.10.11-x64;ubuntu.14.04-x64;ubuntu.16.04-x64;centos.7-x64;rhel.7.2-x64;debian.8-x64;fedora.23-x64;opensuse.13.2-x64</RuntimeIdentifiers>
   </PropertyGroup>
 
   <ItemGroup>
@@ -28,36 +17,16 @@
   </ItemGroup>
 
   <ItemGroup>
-	<PackageReference Include="Microsoft.TemplateEngine.Utils" Version="*" />
-	<PackageReference Include="Microsoft.TemplateEngine.Abstractions" Version="*" />
-	<PackageReference Include="Microsoft.TemplateEngine.Edge" Version="*" />
+    <PackageReference Include="Microsoft.TemplateEngine.Utils" Version="$(PackageVersion)" />
+    <PackageReference Include="Microsoft.TemplateEngine.Abstractions" Version="$(PackageVersion)" />
+    <PackageReference Include="Microsoft.TemplateEngine.Edge" Version="$(PackageVersion)" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Sdk">
-      <Version>1.0.0-alpha-20161104-2</Version>
-      <PrivateAssets>All</PrivateAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.DotNet.Cli.Utils">
-      <Version>1.0.0-preview2-003121</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.CommandLineUtils">
-      <Version>1.0.0</Version>
-    </PackageReference>
-    <PackageReference Include="Newtonsoft.Json">
-      <Version>9.0.1</Version>
-    </PackageReference>
+    <PackageReference Include="Microsoft.NETCore.App" Version="1.0.0" />
+    <PackageReference Include="Microsoft.DotNet.Cli.Utils" Version="1.0.0-preview2-003121" />
+    <PackageReference Include="Microsoft.Extensions.CommandLineUtils" Version="1.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp1.0' ">
-    <PackageReference Include="Microsoft.NETCore.App">
-      <Version>1.0.0</Version>
-    </PackageReference>
-  </ItemGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <DefineConstants>$(DefineConstants);RELEASE</DefineConstants>
-  </PropertyGroup>
-  
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
