+++ test/Microsoft.TemplateEngine.Core.UnitTests/ConditionalTests.CStyleEvaluator.cs
@@ -1,0 +1,1530 @@
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Text;
+using Microsoft.TemplateEngine.Core.Contracts;
+using Xunit;
+
+namespace Microsoft.TemplateEngine.Core.UnitTests
+{
+    public partial class ConditionalTests
+    {
+        #region Original Tests
+
+        [Fact]
+        public void VerifyIfEndifTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = true };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = true };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueConditionContainsTabs()
+        {
+            string value = @"Hello
+    #if " + "\t" + @" (VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = true };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueConditionQuotedString()
+        {
+            string value = @"Hello
+    #if (""Hello" + "\t" + @"There"" == VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = "Hello\tThere" };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueConditionLiteralFirst()
+        {
+            string value = @"Hello
+    #if (3 > VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueConditionLiteralAgainst()
+        {
+            string value = @"Hello
+    #if(3 > VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifTrueConditionAgainstIf()
+        {
+            string value = @"Hello
+    #if(VALUE)
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = true };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifFalseCondition()
+        {
+            string value = @"Hello
+    #if VALUE
+value
+    #else
+other
+    #endif
+There";
+            string expected = @"Hello
+other
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection { ["VALUE"] = false };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifEndifTrueFalseCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifEndifTrueTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifEndifFalseTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #endif
+There";
+            string expected = @"Hello
+other
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseEndifTrueFalseCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #else
+other2
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseEndifFalseTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #else
+other2
+    #endif
+There";
+            string expected = @"Hello
+other
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseEndifFalseFalseCondition()
+        {
+            string value = @"Hello
+    #if VALUE
+value
+    #elseif VALUE2
+other
+    #else
+other2
+    #endif
+There";
+            string expected = @"Hello
+other2
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyNestedIfTrueTrue()
+        {
+            string value = @"Hello
+    #if (VALUE)
+        #if (VALUE2)
+value
+        #else
+other
+        #endif
+    #else
+other2
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifElseEndifTrueTrueCondition()
+        {
+            string value = @"Hello
+        #if (VALUE)
+value
+        #elseif (VALUE2)
+other
+        #else
+other2
+        #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifElseEndifTrueFalseCondition()
+        {
+            string value = @"Hello
+        #if (VALUE)
+value
+        #elseif (VALUE2)
+other
+        #else
+other2
+        #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifElseEndifFalseTrueCondition()
+        {
+            string value = @"Hello
+        #if (VALUE)
+value
+        #elseif (VALUE2)
+other
+        #else
+other2
+        #endif
+There";
+            string expected = @"Hello
+other
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifElseEndifFalseFalseCondition()
+        {
+            string value = @"Hello
+        #if VALUE
+value
+        #elseif VALUE2
+other
+        #else
+other2
+        #endif
+There";
+            string expected = @"Hello
+other2
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifEndifTrueFalseFalseCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #elseif (VALUE3)
+other2
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true,
+                ["VALUE2"] = false,
+                ["VALUE3"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifEndifFalseTrueFalseCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #elseif (VALUE3)
+other2
+    #endif
+There";
+            string expected = @"Hello
+other
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = true,
+                ["VALUE3"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseifElseifEndifFalseFalseTrueCondition()
+        {
+            string value = @"Hello
+    #if (VALUE)
+value
+    #elseif (VALUE2)
+other
+    #elseif (VALUE3)
+other2
+    #endif
+There";
+            string expected = @"Hello
+other2
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false,
+                ["VALUE2"] = false,
+                ["VALUE3"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueEqualsCondition()
+        {
+            string value = @"Hello
+    #if (VALUE == 2)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2L
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueNotEqualsCondition()
+        {
+            string value = @"Hello
+    #if (VALUE != 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 4
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueGreaterThanCondition()
+        {
+            string value = @"Hello
+    #if (VALUE > 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 4
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifOperandStealing()
+        {
+            string value = @"Hello
+    #if ((VALUE == 3) == true)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 3L
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifOperandStealing2()
+        {
+            string value = @"Hello
+    #if (!VALUE == true)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueGreaterThanOrEqualToCondition()
+        {
+            string value = @"Hello
+    #if (VALUE >= 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 3
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifFalseGreaterThanOrEqualToCondition()
+        {
+            string value = @"Hello
+    #if (VALUE >= 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueLessThanCondition()
+        {
+            string value = @"Hello
+    #if (VALUE < 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueLessThanOrEqualToCondition()
+        {
+            string value = @"Hello
+    #if (VALUE <= 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 3
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueNotCondition()
+        {
+            string value = @"Hello
+    #if (!VALUE)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = false
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueNotNotCondition()
+        {
+            string value = @"Hello
+    #if (!!VALUE)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = true
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueAndCondition()
+        {
+            string value = @"Hello
+    #if (VALUE < 3 && VALUE > 0)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueXorCondition()
+        {
+            string value = @"Hello
+    #if (VALUE < 3 ^ VALUE == 7)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueAndAndCondition()
+        {
+            string value = @"Hello
+    #if (VALUE < 3 && VALUE < 4 && VALUE < 5)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueOrCondition()
+        {
+            string value = @"Hello
+    #if (VALUE == 6 || VALUE < 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueOrOrCondition()
+        {
+            string value = @"Hello
+    #if (VALUE == 6 || VALUE == 7 || VALUE < 3)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueOrAndCondition()
+        {
+            string value = @"Hello
+    #if (VALUE == 6 || (VALUE != 7 && VALUE < 3))
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueAndOrCondition()
+        {
+            string value = @"Hello
+    #if ((VALUE != 7 && VALUE < 3) || VALUE == 6)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueBitwiseAndEqualsCondition()
+        {
+            string value = @"Hello
+    #if ((VALUE & 0xFFFF) == 2)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueBitwiseOrEqualsCondition()
+        {
+            string value = @"Hello
+    #if (VALUE | 0xFFFD == 0xFFFF)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueShlCondition()
+        {
+            string value = @"Hello
+    #if (VALUE << 1 == 8)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 4
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueShrCondition()
+        {
+            string value = @"Hello
+    #if (VALUE >> 1 == 2)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 4
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfEndifTrueGroupedCondition()
+        {
+            string value = @"Hello
+    #if ((VALUE == 2) && (VALUE2 == 3))
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection
+            {
+                ["VALUE"] = 2L,
+                ["VALUE2"] = 3L
+            };
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifConditionUsesNull()
+        {
+            string value = @"Hello
+    #if (VALUE2 == null)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifConditionUsesFalse()
+        {
+            string value = @"Hello
+    #if (!false)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifConditionUsesDouble()
+        {
+            string value = @"Hello
+    #if (1.2 < 2.5)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfElseEndifConditionUsesFalsePositiveHex()
+        {
+            string value = @"Hello
+    #if (0xChicken == null)
+value
+    #endif
+There";
+            string expected = @"Hello
+value
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyIfNoCondition()
+        {
+            string value = @"Hello
+    #if
+value
+    #endif
+There";
+            string expected = @"Hello
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyConditionAtEnd()
+        {
+            string value = @"Hello
+    #if (1.2 < 2.5)";
+            string expected = @"Hello
+";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyExcludeNestedCondition()
+        {
+            string value = @"Hello
+    #if false
+        #if true
+            #if true
+            #endif
+        #endif
+        #if true
+            #if true
+            #endif
+        #endif
+    #endif";
+            string expected = @"Hello
+";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        // Note: In the long-term, the expected result would be @"Hello"
+        // without the newline in the existing test. 
+        // In general, we'd walk back from the if (which must be the first non-whitespace on a line)
+        // walking back up to and including the previous newline. 
+        // (the output stream may get unwound a bit to do this).
+        // So overall:
+        //  - we never take the newline just before the start of a conditional
+        //  - we always take the final newline of a conditional block
+        //
+        // With this in mind, numerous tests probably need some minor adjustments.
+        [Fact]
+        public void VerifyExcludeNestedConditionInNonTakenBranch()
+        {
+            string value = @"Hello
+    #if true
+    #else
+        #if true
+            #if true
+            #endif
+        #endif
+        #if true
+            #if true
+            #endif
+        #endif
+    #endif
+There";
+            string expected = @"Hello
+There";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            bool changed = processor.Run(input, output, 28);
+            Verify(Encoding.UTF8, output, changed, value, expected);
+        }
+
+        [Fact]
+        public void VerifyEmitStrayToken()
+        {
+            string value = @"Hello
+    #endif
+    #else
+    #elseif foo";
+            string expected = @"Hello
+    #endif
+    #else
+    #elseif foo";
+
+            byte[] valueBytes = Encoding.UTF8.GetBytes(value);
+            MemoryStream input = new MemoryStream(valueBytes);
+            MemoryStream output = new MemoryStream();
+
+            VariableCollection vc = new VariableCollection();
+            IProcessor processor = SetupCStyleNoCommentsProcessor(vc);
+
+            //Changes should be made
+            processor.Run(input, output, 28);
+            //Override the change indication - the stream was technically mutated in this case,
+            //  pretend it's false because the inputs and outputs are the same
+            Verify(Encoding.UTF8, output, false, value, expected);
+        }
+
+        #endregion Original Tests
+    }
+}
