+++ version.props
@@ -1,0 +1,5 @@
+<Project ToolsVersion="15.0">
+  <PropertyGroup>
+    <PackageVersion Condition=" '$(PackageVersion)' == '' ">1.0.0-beta1</PackageVersion>
+  </PropertyGroup>
+</Project>
